Add files via upload
This is a matlab code of CDF bias calibration of GPM-IMERG hourly rainfall data. The hourly rain gauge observation data was stored in hrain0 mat file. The hourly GPM-IMERG data was stored in hrain1 mat file. This code was used to adjust the cumulative distribution curve of GPM-IMERG to that of rain gauge. The CDF calibrated hourly GPM-IMERG data was stored in hrain2 variable.